+++ crates/aiken-project/src/blueprint/validator.rs
@@ -11,7 +11,7 @@ use aiken_lang::{
     gen_uplc::CodeGenerator,
     plutus_version::PlutusVersion,
     source_map::SourceMap,
-    tipo::{Type, collapse_links},
+    tipo::Type,
 };
 use miette::NamedSource;
 use serde;
@@ -249,7 +249,7 @@ impl Validator {
 }
 
 pub fn tipo_or_annotation<'a>(module: &'a CheckedModule, arg: &'a TypedArg) -> &'a Type {
-    match collapse_links(arg.tipo.clone()).borrow() {
+    match Type::collapse_links(arg.tipo.clone()).borrow() {
         Type::App {
             module: module_name,
             name: type_name,