+++ crates/aiken/src/cmd/check.rs
@@ -1,10 +1,12 @@
+#![allow(clippy::doc_overindented_list_items)]
+
 use super::build::{trace_filter_parser, trace_level_parser};
 use aiken_lang::{
     ast::{TraceLevel, Tracing},
     test_framework::PropertyTest,
 };
 use aiken_project::{
-    telemetry::json_schema,
+    telemetry::{CoverageMode, json_schema},
     watch::{self, watch_project, with_project},
 };
 use rand::prelude::*;
@@ -15,6 +17,7 @@ use std::{
 };
 
 #[derive(clap::Args)]
+#[clap(disable_version_flag(true))]
 #[command(
     verbatim_doc_comment,
     about = color_print::cstr!(r#"
@@ -63,6 +66,25 @@ pub struct Args {
     /// Maximum number of successful test run for considering a property-based test valid.
     #[clap(long, default_value_t = PropertyTest::DEFAULT_MAX_SUCCESS, value_name="UINT")]
     max_success: usize,
+
+    /// Display options for the property tests labels coverage.
+    ///
+    ///   - relative-to-labels:
+    ///       Ratio of each label over the total number of labels.
+    ///       Better when labels are mutually-exclusive and present in each test.
+    ///
+    ///   - relative-to-tests:
+    ///       Ratio of each label over the total number of tests/iterations.
+    ///       Better when labels are occasional and non-exclusive per test.
+    #[clap(
+        long,
+        short = 'P',
+        default_value_t = CoverageMode::default(),
+        value_parser = CoverageMode::parse,
+        value_name="COVERAGE_MODE",
+        verbatim_doc_comment
+    )]
+    property_coverage: CoverageMode,
 
     /// Only run tests if they match any of these strings.
     /// You can match a module with `-m aiken/list` or `-m list`.
@@ -117,6 +139,7 @@ pub fn exec(
         debug,
         show_json_schema,
         match_tests,
+        property_coverage,
         exact_match,
         watch,
         trace_filter,
@@ -144,6 +167,7 @@ pub fn exec(
                 exact_match,
                 seed,
                 max_success,
+                property_coverage,
                 match trace_filter {
                     Some(trace_filter) => trace_filter(trace_level),
                     None => Tracing::All(trace_level),
@@ -165,6 +189,7 @@ pub fn exec(
                     exact_match,
                     seed,
                     max_success,
+                    property_coverage,
                     match trace_filter {
                         Some(trace_filter) => trace_filter(trace_level),
                         None => Tracing::All(trace_level),